Move all tests into subfolder of each app
