+++ utilities/tests/test_holiday.py
@@ -3,21 +3,7 @@ import datetime
 from django.db import IntegrityError
 from django.test import TestCase
 
-from .models import Holiday, Utility
-
-
-class UtilityModelTests(TestCase):
-    def test_create_and_str(self):
-        """Test creating a utility and its string representation."""
-        utility = Utility.objects.create(name="PG&E", timezone="America/Los_Angeles")
-        self.assertIsNotNone(utility.pk)
-        self.assertEqual(str(utility), "PG&E")
-
-    def test_name_uniqueness(self):
-        """Test that utility names must be unique."""
-        Utility.objects.create(name="PG&E", timezone="America/Los_Angeles")
-        with self.assertRaises(IntegrityError):
-            Utility.objects.create(name="PG&E", timezone="America/New_York")
+from utilities.models import Holiday, Utility
 
 
 class HolidayModelTests(TestCase):
